+++ Classes/Utility/AbstractUtility.php
@@ -2,8 +2,7 @@
 namespace T3v\T3vCore\Utility;
 
 use TYPO3\CMS\Core\SingletonInterface;
-use TYPO3\CMS\Core\Utility\GeneralUtility;
-use TYPO3\CMS\Extbase\Object\ObjectManager;
+use TYPO3\CMS\Extbase\Object\ObjectManagerInterface;
 
 /**
  * The abstract utility class.
@@ -14,14 +13,16 @@ abstract class AbstractUtility implements SingletonInterface {
   /**
    * The object manager.
    *
-   * @var \TYPO3\CMS\Extbase\Object\ObjectManager
+   * @var \TYPO3\CMS\Extbase\Object\ObjectManagerInterface
    */
   protected $objectManager;
 
   /**
-   * Constructs a new abstract utility.
+   * Injects the object manager.
+   *
+   * @param \TYPO3\CMS\Extbase\Object\ObjectManagerInterface $objectManager
    */
-  public function __construct() {
-    $this->objectManager = GeneralUtility::makeInstance(ObjectManager::class);
+  public function injectObjectManager(ObjectManagerInterface $objectManager) {
+    $this->objectManager = $objectManager;
   }
 }